status: Watch clusteroperators
Add a watch on clusteroperators in the status controller so that it updates
or recreates the ingress clusteroperator as necessary should something else
update or delete it.

This commit fixes bug 2093462.

https://bugzilla.redhat.com/show_bug.cgi?id=2093462

* manifests/00-cluster-role.yaml: Allow the operator to list and watch
clusteroperators.
* pkg/manifests/bindata.go: Regenerate.
* pkg/operator/controller/status/controller.go (New): Watch
clusteroperators with a map function and predicate to reconcile the default
ingresscontroller if clusteroperators.config.openshift.io/ingress is
updated.
* test/e2e/all_test.go (TestAll): Add the new
TestOperatorRecreatesItsClusterOperator test to the serial tests.
* test/e2e/operator_test.go (TestOperatorRecreatesItsClusterOperator): New
test.  Verify that the operator recreates the "ingress" clusteroperator if
the clusteroperator is deleted.