ingressclass: Fix API group name in watch predicate
Use the correct group name in the predicate for the watch on ingressclasses
in the ingressclass controller so that the controller reconciles events for
ingressclasses that reference the ingress-to-route controller.

This commit fixes bug 1949435.

https://bugzilla.redhat.com/show_bug.cgi?id=1949435

* pkg/operator/controller/ingressclass/controller.go
(ingressClassHasIngressController): Use the correct string for the group
name.
* test/e2e/operator_test.go (TestDefaultIngressClass): Verify that the
ingressclass controller recreates the ingressclass should it be deleted.