Allow load-balancer-source-ranges annotation
Ignore the "service.beta.kubernetes.io/load-balancer-source-ranges" service
annotation when computing the Upgradeable status condition so that users
can continue to use the annotation with OpenShift 4.8.

Setting the spec.loadBalancerSourceRanges field still affects the
Upgradeable status condition.  This means that users who have been setting
spec.loadBalancerSourceRanges will be prompted to migrate to using the
"service.beta.kubernetes.io/load-balancer-source-ranges" annotation on
OpenShift 4.8 before upgrading to OpenShift 4.9.  In OpenShift 4.9, we will
add an API field to IngressController for configuring source ranges and set
Upgradeable=False if the annotation is set to prompt users to switch to the
new API field, so the user will be able to migrate from the annotation to
this API after upgrading to OpenShift 4.9.

* pkg/operator/controller/ingress/load_balancer_service.go
(loadBalancerServiceExternallyModified): Ignore the
"service.beta.kubernetes.io/load-balancer-source-ranges" annotation.
* pkg/operator/controller/ingress/load_balancer_service_test.go
(TestLoadBalancerServiceExternallyModified):
* pkg/operator/controller/ingress/status_test.go
(TestComputeIngressUpgradeableCondition): Expect the
"service.beta.kubernetes.io/load-balancer-source-ranges" annotation to be
ignored and not to affect the Upgradeable status condition.