ingressclass: Handle ingresscontroller deletion
If the ingresscontroller corresponding to an ingressclass does not
exist (for example, because the ingresscontroller was deleted), delete the
ingressclass.

This commit fixes bug 1950978.

https://bugzilla.redhat.com/show_bug.cgi?id=1950978

* manifests/00-cluster-role.yaml: Allow the operator to delete
ingressclasses.
* pkg/manifests/bindata.go: Regenerate.
* pkg/operator/controller/ingressclass/controller.go (Reconcile): Pass the
ingresscontroller's namespaced name to ensureIngressClass.
* pkg/operator/controller/ingressclass/ingressclass.go
(ensureIngressClass): Change ingresscontroller name parameter type from
string to namespaced-name.  Check if the ingresscontroller exists and is
not marked for deletion, and pass a corresponding Boolean value to
desiredIngressClass.
(desiredIngressClass): Return false, nil if the ingresscontroller does not
exist.
* test/e2e/operator_test.go (TestCustomIngressClass): New test.  Verify
that the ingressclass controller creates an ingressclass for a custom
ingresscontroller and deletes the ingressclass when the ingresscontroller
is deleted.