test/e2e: TestHTTPHeaderBufferSize must wait for old pod to delete
The addition of minReadySeconds slows the rate at which config changes
rollout - the old logic in the test made an update to the IC and assumed
the new pod being ready == high likelihood the new pod and router config
was live, but instead the old pod was actually just as likely to get
the request.

Wait until the old pod is gone before continuing.