BZ https://bugzilla.redhat.com/1901648 The following use case is fixed now by this fix:
1. Create Ingress with a custom domain
2. Ingress status gets updated by OpenShift Ingress controller with router canonical hostname
3. Use external-dns to sync with Route 53

The problem was the canonical router hostname didn't exist in the DNS. It is not created by OpenShift. OpenShift creates this *.apps.<cluster_name>.<base_domain> DNS record and not this one apps.<cluster_name>.<base_domain>. So canonical router hostname was not right. Now this fix sets it to router-default.apps.<cluster_name>.<base_domain>