canary: Add priority class to canary daemonset
OpenShift components are required to specify a
PriorityClassName for daemonsets/deployments.

This commit ensures that the Canary daemonset has the
`system-cluster-critical` priority class set.

assets/canary/daemonset.yaml:
Set the `priorityClassName` field to be `system-cluster-critical`.

pkg/operator/controller/canary/daemonset.go:
Update the Canary daemonset when `priorityClassName` changes.

pkg/operator/controller/canary/daemonset_test.go:
Add a test to ensure that `priorityClassName` is set correctly by
`desiredCanaryDaemonSet`.
Add a test to verify that the Canary daemonset is updated
when `priorityClassName` is mutated.

This commit is in support of Bug 1954892.