Add unsupported config override for maxconn
Add an unsupported config override for setting HAProxy's global maxconn
setting.  If the config override is unspecified, the current default of
20000 is used.  If the config override has the value "-1", the maxconn
setting is omitted so that HAProxy automatically detects a reasonable value
based on the value of ulimit -n.

* pkg/operator/controller/ingress/deployment.go
(RouterMaxConnectionsEnvName): New const.
(desiredRouterDeployment): Add unsupported config override for
ROUTER_MAX_CONNECTIONS.
* pkg/operator/controller/ingress/deployment_test.go
(TestDesiredRouterDeployment): Verify that desiredRouterDeployment sets
ROUTER_MAX_CONNECTIONS as expected.
* test/e2e/operator_test.go (TestMaxConnectionsUnsupportedConfigOverride):
Verify that the operator sets ROUTER_MAX_CONNECTIONS to the appropriate
value based on the value specified in the unsupported configuration
override if an override is specified.