crl: Watch client CA configmap in operand namespace
Change the crl controller watch on configmaps to watch for updates to the
client CA configmap that is in the "openshift-ingress" operand namespace.

Before this commit, the crl controller watched the user-managed client CA
configmap in the "openshift-config" global config namespace.  However, the
crl controller builds the CRL configmap from the operator-managed client CA
configmap in the operand namespace, so it makes sense to watch this
configmap instead.

This commit is related to bug 1989058.

https://bugzilla.redhat.com/show_bug.cgi?id=1989058

* pkg/operator/controller/crl/controller.go (New): Use the name of the
operator-managed configmap when building the index.
(configmapToIngressController): Ignore the configmap if it isn't in the
operand namespace.