crl: Watch CRL configmaps
Add a watch on CRL configmaps so that the controller recreates a configmap
if the user deletes it.

In principle, the same watch could be used for both client CA configmaps
and CRL configmaps, but they are using separate indices and perform
logically distinct functions, so it makes sense to use separate watches.
The watches share an informer cache, so the overhead of using two watches
instead of one is small.

This commit fixes bug 1989058.

https://bugzilla.redhat.com/show_bug.cgi?id=1989058

* pkg/operator/controller/crl/controller.go (crlConfigmapIndexFieldName):
New const.
(New): Add an index with name crlConfigmapIndexFieldName for
ingresscontrollers keyed by CRL configmap names.  Add a watch for
ingresscontrollers with mapfunc crlConfigmapToIngressController.
(ingressControllersWithConfigmap): Rename...
(ingressControllersWithClientCAConfigmap): ...to this.
(configmapToIngressController): Rename...
(clientCAConfigmapToIngressController): ...to this.
(ingressControllersWithCRLConfigmap): New method.  Return a slice of all
ingresscontrollers that reference the specified CRL configmap.
(crlConfigmapToIngressController): New method.  Map a CRL configmap to
a slice of reconcile requests, using ingressControllersWithCRLConfigmap.