Set ROUTER_DOMAIN to enable route subdomain field
Set ROUTER_DOMAIN to the IngressController's domain in the router
deployment so that the router can use the domain to construct a default
host for routes that specify an empty value for spec.host and a nonempty
value for spec.subdomain.

* pkg/operator/controller/ingress/deployment.go (desiredRouterDeployment):
Set ROUTER_DOMAIN.
* pkg/operator/controller/ingress/deployment_test.go
(TestDesiredRouterDeployment): Verify that ROUTER_DOMAIN is not set.
(TestDesiredRouterDeploymentSpecAndNetwork): Verify that ROUTER_DOMAIN is
set to the expected value.