setDefaultPublishingStrategy: Reformat with switch
Refactor the update logic in setDefaultPublishingStrategy.  Also, fix
setDefaultPublishingStrategy to return true if the scope changed.  (Nothing
uses this return value, so this change is only for correctness.)

* pkg/operator/controller/ingress/controller.go
(setDefaultPublishingStrategy): Use a switch statement for the update logic
so that the logic only looks at parameters related to the selected endpoint
publishing strategy type.