Support "auto-delete-load-balancer" annotation
Support the "ingress.operator.openshift.io/auto-delete-load-balancer"
ingresscontroller annotation.  If this annotation is set on an
ingresscontroller, the ingresscontroller's scope has been changed, and the
current platform requires deleting and recreating a service load-balancer
to change its scope, then the operator deletes and recreates the
LoadBalancer service for the ingresscontroller automatically.

* pkg/operator/controller/ingress/load_balancer_service.go
(autoDeleteLoadBalancerAnnotation): New const.
(platformsWithMutableScope): New variable.  Define the set of platforms
that support mutating load-balancer scope *without* deleting and recreating
a service load-balancer.
(managedLoadBalancerServiceAnnotations): Use platformsWithMutableScope.
(ensureLoadBalancerService): Check for the "auto-delete-load-balancer"
annotation on the ingresscontroller and pass an argument to
updateLoadBalancerService to indicate the annotation's presence.
(updateLoadBalancerService): Check whether scope changed, the annotation is
set, and the platform requires deleting and recreating the service to
change scope, and do so if these conditions are all met.
(scopeEqual): New function.  Return true if the two given services have the
same scope, false otherwise.
* test/e2e/operator_test.go (TestScopeChange): Add steps to verify that the
"auto-delete-load-balancer" annotation works properly on platforms that
require it.