TestIngressControllerScale: Use a private IC
Use a custom, private ingresscontroller for the scale test, and scale up
from 1 replica to 2 and then back down to 1 replica, to avoid scheduling
problems when creating an additional router pod replica during scale-up.

Before this change, the test used the default ingresscontroller and
attempted to scale it up from 2 replicas to 3, which could fail if
insufficient worker nodes were available.  This change reduces the
likelihood of failures by using a private ingresscontroller (which doesn't
require host ports) and scaling from 1 replica to 2 (which doesn't require
as many worker nodes).

* test/e2e/operator_test.go (TestIngressControllerScale): Use a custom,
private ingresscontroller, and start with 1 replica.