Azure: Add Support for Azure Stack Hub
Azure Stack is on-prem Azure, so the ARM endpoint, which is
known in advance for public Azure or Government Azure, must be
provided through configuration. This commit adds support for Azure
Stack by checking the infrastructure object to determine the cloud
and in the case of Azure Stack, uses the provided ARMEndpoint.