Fix update retry logic in TestConfigurableRoute*
Fix intermittent failures in the TestConfigurableRouteRBAC,
TestConfigurableRouteNoSecretNoRBAC, and
TestConfigurableRouteNoConsumingUserNoRBAC tests that were caused by
incorrect update retry logic.

These tests use the eventuallyUpdateIngressSpec and
eventuallyUpdateIngressStatus helper functions to update the cluster
ingress config.  Before this commit, these helper functions attempted to
update the ingress config using a polling loop that gets the current
config, assigns a value provided as an argument to the helper function to a
field in the config value from the API, and then updates the config in the
API using the modified config value.  The argument value is a complex data
type, and so the assignment causes the argument value and the modified
config value to reference the same memory.  If the update failed, the next
iteration of the polling loop would do another get from the API and
overwrite the argument value.  As a result, the assignment and update in
the second and subsequent iterations of the polling loop would use the
value from the API, not the value originally provided in the argument.

This commit replaces the assignment in the polling loop with a deep copy so
that the original argument value is not overwritten by the get.  As a
result, the test should be more resilient when updates fail.

Follow-up to commit baf2d3e1c1fea1a3c9f40dcc44c5d7418b347f75.

* test/e2e/configurable_route_test.go (eventuallyUpdateIngressSpec)
(eventuallyUpdateIngressStatus): Use DeepCopyInto instead of an assignment.