desiredRouterDeployment: Default 1 replica on SNO
When determining the default number of router pod replicas for an
ingresscontroller that does not specify replicas, check the infrastructure
topology, and default to 1 replica if the topology is "SingleReplica"
(i.e., Single-Node OpenShift).

When the publishing strategy is "Internal", the installer generates a
default ingresscontroller manifest, and this manifest does not specify
replicas.  As a result, the ingress operator used the default value of 2
replicas.  On single-node clusters, this caused the ingresscontroller to
reported degraded status because only 1 pod replica can be scheduled per
ingresscontroller per node.

This issue does not arise when the publishing strategy is "External"
because in this case, the installer does not generate a default
ingresscontroller manifest; instead, the ingress operator creates the
default ingresscontroller, and the ingress operator does check the topology
and specify an appropriate, explicit number of replicas in the
ingresscontroller that it creates.

Follow-up to commit 4abef0342262a8ab644a5d817574c4a3b09b3fc6.

This commit fixes bug 2005351.

https://bugzilla.redhat.com/show_bug.cgi?id=2005351

* pkg/operator/controller/ingress/deployment.go (ensureRouterDeployment):
Pass infraConfig to desiredRouterDeployment.
(desiredRouterDeployment): Add a parameter for the infrastructure config.
Use the infrastructure to check whether the infrastructure topology.  If
the topology is "SingleReplica" and replicas is not specified, default to 1
replica.
* pkg/operator/controller/ingress/deployment_test.go
(TestDesiredRouterDeployment): Verify that desiredRouterDeployment behaves
correctly when the infrastructure topology is "SingleReplica".