Use externalTrafficPolicy: Cluster with OVN
OVN-Kubernetes in OpenShift 4.9 and earlier does not support
externalTrafficPolicy: Local, and specifying it is reported to result in
imbalanced traffic for some users.  This commit checks the cluster network
type and sets externalTrafficPolicy: Cluster on the LoadBalancer-type
service that the operator creates for ingress if the network type is
"OVN-Kubernetes".

This commit fixes bug 2079517.

https://bugzilla.redhat.com/show_bug.cgi?id=2079517

* pkg/operator/controller/ingress/controller.go (ensureIngressController):
Pass the network config to ensureLoadBalancerService and
syncIngressControllerStatus.
* pkg/operator/controller/ingress/load_balancer_service.go
(ensureLoadBalancerService, loadBalancerServiceIsUpgradeable):
Add a parameter for the network config.  Pass the network type
from the network config to desiredLoadBalancerService.
(desiredLoadBalancerService): Add a parameter for the network type.  Set
externalTrafficPolicy: Cluster if the network type is "OVN-Kubernetes".
* pkg/operator/controller/ingress/load_balancer_service_test.go
(TestDesiredLoadBalancerService): Add a test case for OVN-Kubernetes.
* pkg/operator/controller/ingress/status.go (syncIngressControllerStatus):
Add a parameter for the network config.  Pass the config to
computeIngressUpgradeableCondition.
(computeIngressUpgradeableCondition): Add a parameter for the network
config.  Pass the network config to loadBalancerServiceIsUpgradeable.
* pkg/operator/controller/ingress/status_test.go
(TestComputeIngressUpgradeableCondition): Pass the network type to
desiredLoadBalancerService and the network config to
computeIngressUpgradeableCondition.