NE825: Update router to use random balancing algorithm once again
After attempting to use random as the default algorithm and subsequently
reverting, we discovered that reducing haproxy's weight to 1 for backends
with a single service significantly reduces the memory overhead of
the random balance algorithm and provides the same exactly functionality.

We are now turning back on random again since we have mitigated the memory
overhead issue.