Determine number of replicas according to `DefaultPlacement`
Implementation of the https://github.com/openshift/enhancements/blob/master/enhancements/single-node/single-node-openshift-with-workers.md
enhancement.

The installation process now sets the new `DefaultPlacement`
API field in the Ingress Config CR's status (openshift/installer#5746
still unmerged at the time of writing this message).

This commit introduced two new changes -

- While creating the default `IngressController`, or while creating
deployments corresponding to `IngressController` resources which
don't have their number of replicas set, we determine the number
of replicas according to the value of the new `DefaultPlacement`
parameter.

- When creating Deployments out of `IngressController` resources
which don't have their nodeSelector set, we determine the default
nodeSelector according to the new `DefaultPlacement` API field.

See the enhancement, new API documentation, implementation and code
comments in the diff for the exact details on how, why and when we
set the fields and to which values.