Allow users to tune kubelet probe timeouts
Allow users to set the timeoutSeconds parameter on the liveness, readiness,
and startup probes for router deployments other than that of the default
ingresscontroller.

* pkg/operator/controller/ingress/deployment.go (ensureRouterDeployment):
Check the ingresscontroller's name to determine whether it is the default
ingresscontroller and pass a Boolean value indicating the same to
desiredRouterDeployment and updateRouterDeployment.
(desiredRouterDeployment): Add a defaultIngressController parameter and
pass its argument to deploymentTemplateHash.
(deploymentHash, deploymentTemplateHash): Add a defaultIngressController
parameter and pass the argument to hashableDeployment.
(hashableDeployment): Add a defaultIngressController parameter and pass its
argument to hashableProbe.
(hashableProbe): Add a defaultIngressController parameter.  If its argument
is false, don't include the probe's timeoutSeconds parameter in the
hashable probe.
(updateRouterDeployment): Add a defaultIngressController parameter and pass
its argument to deploymentHash.
(deploymentConfigChanged): Add a defaultIngressController parameter and use
it along with the new copyProbe function to set the updated deployment's
probe parameters.
(copyProbe): New function.  Copy one probe to another, preserving
timeoutSeconds if the defaultIngressController parameter is false so as to
allow users to modify the timeoutSeconds parameter.
* pkg/operator/controller/ingress/deployment_test.go
(checkDeploymentHasContainer, TestDesiredRouterDeployment)
(TestDeploymentHash): Pass a false defaultIngressController argument to
functions that need it.
(TestDeploymentConfigChanged): Add test cases to verify that
deploymentConfigChanged detects changes to timeoutSeconds and other probe
parameters for the default ingresscontroller and ignores changes to
timeoutSeconds for other ingresscontrollers.
* test/e2e/operator_test.go
(TestTunableRouterKubeletProbesForDefaultIngressController)
(TestTunableRouterKubeletProbesForCustomIngressController): New tests.
* test/e2e/util.go (cmpProbes, probe): New helpers that are used by the new
tests.