Remove special-casing for default IC probe params
Remove logic that special-cases handling of the default ingresscontroller.
This special-casing prevented the user from modifying the kubelet probe
timeouts of the default ingresscontroller while allowing the user to modify
the kubelet probe timeouts for custom ingresscontrollers.  However, the
special-casing was not part of the enhancement design document, so it
should not be in the implementation.  After this commit, the user can
modify the kubelet probe timeouts for default and custom ingresscontrollers
alike.

This commit fixes bug 2076193.

https://bugzilla.redhat.com/show_bug.cgi?id=2076193

* pkg/operator/controller/ingress/deployment.go (ensureRouterDeployment):
Delete the defaultIngressController variable and argument to
desiredRouterDeployment and updateRouterDeployment.
(desiredRouterDeployment, deploymentHash, deploymentTemplateHash)
(hashableDeployment, hashableProbe, updateRouterDeployment)
(deploymentConfigChanged): Delete the defaultIngressController parameter
and special-casing for the default ingresscontroller.
(hashableProbe): Use copyProbe.
(copyProbe): Copy logic that was previously in hashableProbe.  With this
change, hashableProbe and deploymentConfigChanged both use copyProbe and
thus behave consistently for all probe parameters.
* pkg/operator/controller/ingress/deployment_test.go (checkDeploymentHash):
Delete the defaultIngressController argument to deploymentTemplateHash.
(TestTuningOptions, TestDesiredRouterDeployment)
(TestDesiredRouterDeploymentSpecAndNetwork): Delete the
defaultIngressController argument to desiredRouterDeployment.
(TestDeploymentHash): Delete the defaultIngressController argument to
deploymentHash and deploymentTemplateHash.
(TestDeploymentConfigChanged): Delete test cases for changing probe
parameters for the default ingresscontroller as it is no longer
special-cased.
* test/e2e/operator_test.go
(TestTunableRouterKubeletProbesForDefaultIngressController): Delete test.