desiredRouterDeployment: Fix for port defaulting
Specify deployment.spec.template.spec.containers[0].ports[*].protocol
in router deployments.

Before this commit, the operator left the protocol field unspecified, and
as a consequence, the update logic would keep trying to revert the default
value that the API set for this field.  The default value for this field is
"TCP", so specifying "TCP" has the result that the API does not need to
update the deployment to fill in the default value, and so the operator's
update logic is no longer triggered by the API's defaulting of this field.

Follow-up to commit af653f9fa7368cf124e11b7ea4666bc40e601165.

This commit is related to bug 2081447.

https://bugzilla.redhat.com/show_bug.cgi?id=2081447

* pkg/operator/controller/ingress/deployment.go (desiredRouterDeployment):
Specify the protocol field on container ports.
* pkg/operator/controller/ingress/deployment_test.go
(TestDeploymentConfigChanged): Add test cases to verify that updates that
set container ports' protocol to the default value are ignore and that
updates that change the protocol to a non-default value are not ignored.