Makefile: change test-e2e entrypoint to TestAll()
Previously we ran all tests via the wildcard: TEST='.*'.

The new TestALL e2e function groups tests that can be run in parallel
and those that must run in series via two overarching subtests.

The new TestALL function provides for consistent ordering of all e2e
tests; serial tests will not start until the parallel tests complete.
And those tests that can run in parallel can, and should, run in any
order because, by definition, they are deemed to be self-contained.

But some tests which are run in parallel are sensitive to any other
test that modifies the ingress config object; changes to that object
will cause a rolling update of all ingresscontroller's and some of the
parallel tests do not accommodate such an event. Any test that does
modify the ingress config object are always grouped into the serial
subtests.