test/e2e: be explicit with some test expectations
Explicitly assert tests TestHTTPHeaderCapture and
TestHTTPCookieCapture only find 1 pod.

These two tests were flaking in CI because they sometimes coincided
with tests that modified the ingress config object. When that object
is modified a rolling deployment of all ingresscontroller pods can
take place and these two tests fell foul of the expectation that only
one pod was running. The CI logs show that "the pod does not
exist"--and that's because the pod was in terminating state post the
rolling deployment.