test/e2e: retry update on conflict errors
Some tests, notably in test teardown (i.e., defer blocks) fail hard
because the object that is being updated has been changed since it was
last refreshed.

We use the new update*SpecWithRetryOnConflict() functions that allow
finer grained updates on the Spec itself and, if the update fails with
a conflict error, the sequence of get, mutate, update will be retried.