MGMT-10403: Set the `defaultPlacement` for none-platform SNO clusters installed before 4.11
Added a call to a new `handleSingleNode4Dot11Upgrade` function that sets
the defaultPlacement status in the ingress config CR of none-platform
single node clusters to "ControlPlane" if it's not already set. The
situations in which this value is not set are in clusters upgraded from
<4.11 to >=4.11. When that field is not set, it is assumed by this
operator's controllers to have the value "Workers", which is fine for
most clusters, but not a good default for none-platform single-node
clusters as it causes issues when adding workers to them (and is why
starting with 4.11 freshly installed none-platform have it set to
"ControlPlane"). The purpose of this function is to then "fix" those
older clusters so they behave more like freshly installed >=4.11
clusters. Before doing that we also verify that the cluster truly has
only one node, as we wouldn't want to affect clusters that have workers.
That's because It's likely that such clusters already have a custom
ingress configuration that we wouldn't want to potentially mess up, even
though adding workers was previously unsupported.
If defaultPlacement is unset and the cluster doesn't match the criteria
mentioned above, it is anyway given the explicit "Workers" value to
ensure this function doesn't affect it in the future no matter what
changes (e.g. number of nodes), and to make it more consistent with
freshly installed >=4.11 clusters.

This was originally a [non-goal of the enhancement](https://github.com/openshift/enhancements/blob/c2ceb84d86c030650b7b4092d4ee924b16f43faf/enhancements/single-node/single-node-openshift-with-workers.md?plain=1#L185-L190)
but I don't see any harm in fixing it in this manner, it will simplifiy
the need for manual actions and the need to document the difference
between 4.11 clusters and "freshly installed 4.11 clusters" when it
comes to SNO worker expansion.