loadBalancerServiceAnnotationsChanged: Fix update
Fix loadBalancerServiceAnnotationsChanged to use the provided "annotations"
argument when updating annotations.  Before this commit,
loadBalancerServiceAnnotationsChanged used the annotations argument when
checking whether annotations had changed, but then it used
managedLoadBalancerServiceAnnotations when setting the annotations in the
updated service object.

When checking whether the service is upgradeable for the purpose of setting
the ingresscontroller's "Upgradeable" status condition,
loadBalancerServiceTagsModified calls loadBalancerServiceAnnotationsChanged
with an annotations argument that includes the
"service.beta.kubernetes.io/aws-load-balancer-additional-resource-tags"
annotation key, which is not in managedLoadBalancerServiceAnnotations.  As
a result, loadBalancerServiceAnnotationsChanged would report that the
service was not upgradeable, but then the status condition message would
not show that the "aws-load-balancer-additional-resource-tags" annotation
had changed.

* pkg/operator/controller/ingress/load_balancer_service.go
(loadBalancerServiceAnnotationsChanged): Use the provided annotations
argument instead of using managedLoadBalancerServiceAnnotations when
updating annotations.
* pkg/operator/controller/ingress/load_balancer_service_test.go
(TestLoadBalancerServiceAnnotationsChanged): New test.  Verify that
loadBalancerServiceAnnotationsChanged behaves correctly when an unmanaged
annotation is updated or when a managed annotation is added, updated, or
deleted.