loadBalancerServiceAnnotationsChanged: Fix check
Fix loadBalancerServiceAnnotationsChanged to treat nil and empty
annotations maps as equal.  Before this commit, the comparison erroneously
flagged an empty map and a nil map as unequal.  As a result, the ingress
operator could erroneously report that an ingresscontroller with a service
with no annotations was not upgradeable.

This commit fixes bug 2097555.

https://bugzilla.redhat.com/show_bug.cgi?id=2097555

* pkg/operator/controller/ingress/load_balancer_service.go
(loadBalancerServiceAnnotationsChanged): Use a for loop instead of go-cmp
to check whether any managed annotations have changed.
* pkg/operator/controller/ingress/load_balancer_service_test.go
(TestLoadBalancerServiceAnnotationsChanged): Add test cases for empty/nil
current/expected annotations.