Use ingress config to set default LB type on AWS
The cluster ingress config object can now specify the AWS ELB type that was
specified when installing the cluster. The ingress operator will refer to
this object whenever someone deletes the default ingresscontroller and the
operator has to recreate it.

Before this commit, the operator would always set the default ELB type
(Classic ELB) when creating the default ingresscontroller.

This commit implements NE-975.

https://issues.redhat.com/browse/NE-975

* pkg/operator/ingress/controller.go (Reconcile): Pass the result of
isAdmitted to admit.
(admit): Add an "alreadyAdmitted" parameter to indicate whether the
ingresscontroller has previously been admitted and is being re-admitted.
Pass alreadyAdmitted and ingressConfig to setDefaultPublishingStrategy.
(setDefaultPublishingStrategy): Add parameters for the cluster ingress
config and the ingresscontroller's admitted status.  Refactor some
defaulting logic into the new setDefaultProviderParameters function.
(setDefaultProviderParameters): New function.  Initialize the given
EndpointPublishingStrategy's ProviderParameters field and subfields as
appropriate, using the cluster ingress config to set a default LB type on
AWS if the ingresscontroller is being admitted for the first time.
* pkg/operator/controller/ingress/controller_test.go
(TestSetDefaultPublishingStrategySetsPlatformDefaults): Add test cases to
verify the default LB type that is specified in the cluster ingress config
is used for AWS.
(TestSetDefaultPublishingStrategyHandlesUpdates): Add test cases to verify
that changing the default LB type in the cluster ingress config does not
change the LB type on an already admitted ingresscontroller.
* pkg/operator/operator.go (Start): Get the cluster ingress config and pass
it to ensureDefaultIngressController.
(determineReplicasForDefaultIngressController): Delete function.  This is
no longer needed after some refactoring in ensureDefaultIngressController.
(ensureDefaultIngressController): Add a parameter for the cluster ingress
config.  Use it to determine desired replicas and to determine the default
ELB type on AWS.
* test/e2e/operator_test.go (TestAWSLBTypeDefaulting): New test for LB type
defaulting.
* test/e2e/all_test.go (TestAll): Add new E2E test.

Modified-by: Grant Spence <gspence@redhat.com>
Modified-by: Miciah Dashiel Butler Masters <mmasters@redhat.com>